Add Document and Data Toolbar updates
