+++ src/components/custom/common/add-document/index.ts
@@ -1,0 +1,2 @@
+// src\components\custom\common\add-document\index.ts
+export { AddDocumentDialog } from '@/components/custom/common/add-document/add-document-dialog'
