+++ src/components/custom/completed-documents/data-table-toolbar.tsx
@@ -5,7 +5,7 @@ import { Icons } from '@/components/ui/icons'
 import { Table } from '@tanstack/react-table'
 import { Button } from '@/components/ui/button'
 import { Cross2Icon } from '@radix-ui/react-icons'
-import { AddDocumentButton } from '../common/add-document-button'
+import { AddDocumentButton } from '../common/add-document/add-document-button'
 import { doc_type_samples, doc_classification } from '@/lib/dms/data'
 import { DataTableViewOptions } from '@/components/custom/table/data-table-view-options'
 import { DataTableFacetedFilter } from '@/components/custom/table/data-table-faceted-filter'
