+++ src/components/custom/disptach-documents/data-table-toolbar.tsx
@@ -5,7 +5,7 @@ import { Input } from '@/components/ui/input'
 import { DataTableViewOptions } from '@/components/custom/table/data-table-view-options'
 import { doc_type_samples, doc_status, doc_classification } from '@/lib/dms/data'
 import { DataTableFacetedFilter } from '@/components/custom/table/data-table-faceted-filter'
-import { AddDocumentButton } from '../common/add-document-button'
+import { AddDocumentButton } from '../common/add-document/add-document-button'
 import { Icons } from '@/components/ui/icons'
 
 interface DataTableToolbarProps<TData> {
