+++ src/lib/types/document-action-type.ts
@@ -1,0 +1,2 @@
+// src\lib\types\document-action-type.ts
+export type ActionType = 'Receive' | 'Release' | 'Create'
