+++ src/lib/types/index.ts
@@ -1,6 +1,8 @@
+// src\lib\types\index.ts
 export * from '@/lib/types/code'
 export * from '@/lib/types/filter'
 export * from '@/lib/types/status'
 export * from '@/lib/types/documents'
 export * from '@/lib/types/navigation'
 export * from '@/lib/types/api-document'
+export * from '@/lib/types/document-action-type'
